machine: Remove DeleteResult
Its content is redundant either with the passed-in parameters (Name), or
with the returned error (Error, Success). Moreover it's not really clear
which combinations are valid (can we have Error != nil and Success ==
true at the same time?).
This commit removes this for now in favour of just an 'error' return
value. This can be changed later when needed.