WIP - machine: Simplify StartResult
The removed content is redundant either with the passed-in parameters (Name), or
with the returned error (Error, Success). Moreover it's not really clear
which combinations are valid (can we have Error != nil and Success ==
true at the same time?).
This commit removes this for now, but keep the other fields. This can
be changed later when needed.