Issue #593: Rework certificate expiry checks
This removes duplicated code, and tries to make the expiration tests
easier to read.
This also fixes a bug with the warning when the check based on the
bundle build time fails. Before this commit it could say that the
bundle is going to expire in -5 days. Now this errors out in such
situations.
The commit also adds a link to the documentation explaining how to solve
certificate issues.

https://github.com/code-ready/crc/issues/593