Add 'options' parameter to preflight checks
This allows to get rid of the 'parameterized checks' (ie functions named
similarly to other preflight checks, but taking an argument and returning a closure
using that argument so that we can have distro-dependent or network-mode
dependent) preflight checks.

This commit takes the approach of passing an 'options' parameter with
this information to all preflight checks, which allows to remove the
previously mentioned indirection. The preflight checks which need
distro, network-mode, ... can directly get it from the 'options'
instance.

'options' is an interface which is implemented by commonOptions and
linuxOptions (which derives from commonOptions). The only difference
between the two is that on linux we want the distro to be part of the
options.