Introduce units.Size type
This new type will be used to hold sizes in bytes, and to make
it easier to convert between the various units (GB/GiB/...)
The main goal is to stop using int types for memory sizes and disk
sizes, as we get their unit wrong more often than not.
Memory size in crc is MiB while disk size is GiB, but we regularly mix
up MiB/MB and GiB/GB.

The 'Size' type cannot be transparently converted to an int type, so
this has to be done using New() or ConvertTo(). Both methods require a
'unit' parameter, so this should clear up most of the confusion.
Internally, the 'Size' type stores the size in bytes.