Add code to migrate old config
Prior to crc 1.2x, 'memory' was stored in MiB in the config file, and 'disk-size' was stored in GiB.
After this version, they are stored in bytes. This commit adds code to be able
to convert to the new format. It detects the old format through heuristics,
since the disk size/memory size were stored in GiB/MiB, looking at their int
value in the config file will be a lot smaller than the default value in bytes,
but assuming the value is in GiB/MiB and converting it to bytes should give us
something larger than the default value in bytes. If the values we read from
the config file match this heuristic, then we assume it was stored in the old
format and we convert it.