Store 'memory' and 'disk-size' in units.Size
This removes a long-standing ambiguity we had where it was never clear which
unit these config options were storing.
Now the config file will store these values in bytes, and the user can pass the
values on the command line with a unit "crc config set memory 16GB"

Pre-existing configuration files will need to be migrated, this will be done in
a follow-up commit.