os: Improve FileExists behaviour
At the moment, FileExists returns false if and only if os.Stat()
returns a 'FileNotExist' error. In other error cases, FileExists would
return true, while in reality it cannot know.

This commit inverts this behaviour, FileExists() now only returns true
if it's almost sure that the file really exists. When in doubt, it will
return false. As far as I can tell, the existing users should not be
negatively impacted by this change.