Improve version.IsRelease() implementation
This does not need to be a linker variable, as it will be true if and
only if a bundle is embedded in the binary or the macos install path is
set. Both conditions can already be checked at runtime in go code, so we
don't need an additional isRelease variable.