macos: Restart daemon when setup runs
Before the previous commits, the daemon was always restarted when setup
was run as this is done as part of the forced generation of the daemon
launchd plist.
This is no longer the case, but we still need to restart the daemon
when the crc binary was upgraded. This changes
checkIfDaemonAgentRunning() to always fail, and fixDaemonAgentRunning()
to know how to restart the daemon.
When we have a go client for the daemon API, this can be changed to
querying the daemon version to only restart it when needed.