preflight: Simplify SetTray
A 'Tray' label is used to filter when to run the tray preflight checks.
These checks are only run if the tray is set to autostart AND if we used
an installer.
Now that we have the BuildType:Installer label, we can use it for the
tray checks as well instead of having a special case in SetTray.