preflight: Add DaemonStartup label
It can take 2 values, SocketActivated and Manual. Manual means
the daemon was not socket activated. If it always runs through other
means, this is still considered Manual.

This marks the systemd preflight checks with DaemonStartup:
SocketActivated, which will always match at the moment as linux is
always socket activated (at least for now).

However, this will be useful in the subsequent commits where a preflight
test available on all OSes will only be needed when the daemon is
started 'manually'.