exec: Introduce ExecError
This error will wrap stdout/stderr when an error occurs when trying to
run an external executable. In the future it could also wrap errno.
This will be useful to get rid of the (stdout, stderr, err) that all
crcos.Run* commands return, with the `stderr` return value being unused
most of the time, or being printed/manually wrapped in a new error by
each callers.