preflight: Simplify getChecks methods
The preflights now use the new 'options' interface, we no longer need to
call functions with the args we need and then return a 'func() error'
function wrapping these args. This also means we don't need to pass
bundlePath/preset to the functions returning the list of checks to run,
these values are only needed when we generate the options to pass to our
preflight.

This commit is easier to review with 'git diff -w' or similar (ignore
whitespace changes)