vfkit: Retry multiple times to get kernel args
commit 6af9303feca6b added a call to WaitForSSHConnectivity to
workaround unexplained SSH connectivity issues at stop time when trying
to read the kernel arguments.
It could not be reproduced at the time this new commit was written,
but just this keep a retry just in case.

WaitForSSHConnectivity is replaced with errors.Retry(getKernelArgs()).
Since the reason for the SSH connectivity issues are unknown, it's odd
to assume that once WaitForSSHConnectivity succeeded, SSH won't fail
again. By using Retry(), getting the kernel args should eventually work
even if the SSH connectivity is spotty.