preflight: Split cache and suid admin-helper checks
When using homebrew to install crc, crc-admin-helper will be built and
installed by homebrew. In this scenario, the preflight checks only
verify that crc-admin-helper and vfkit are in the expected locations,
and return an error if not.

However, these same checks also verify crc-admin-helper ownership/setuid
status. Since homebrew won't install setuid binaries, this commit
separates these checks from the 'is it cached?' ones to make
crc-admin-helper setuid if needed.