build: Use 'go generate' for compile time parameters
For macOS brew builds, I need a way to specify the path to the vfkit and
crc-admin-helper binaries. This commit adds a way to pass a json file at
build time, which will be used to autogenerate the
pkg/version/helpers_generated.go file.
This file contains a binary name/binary path list.

The path to the json file specified through the env var
CRC_BUILD_PARAMS_PATH.
This json file has the following format:

$ cat build-params.json
[
        {
                "name": "crc-admin-helper",
                "path": "/usr/libexec/crc-admin-helper"
        },
        {
                "name": "vfkit",
                "path": "/usr/bin/vfkit"
        }
]

See
https://blog.carlmjohnson.net/post/2016-11-27-how-to-use-go-generate/
for more details about this approach of using go generate for build time
configuration