Use gvproxy for port exposure
This adjust crc's code to use gvproxy REST API to expose ports.
The URL path it's exposed at is slightly different, there is no
`/network` prefix.

The port exposure can only happen after the VM has started as podman
machine starts gvproxy during the VM startup. This is slightly racy
if we need these ports to be available during early boot, maybe we'll
need to improve podman machine/gvproxy to avoid this.

Port unexposure is not needed anymore as the gvproxy process will be
stopped at the same time as the VM, the ports are no longer kept open by
a long lived process (crc daemon).